refactored constants and util functions to their own files
