+++ ibr/ibr-sdk/src/Visualization.js
@@ -1,18 +1,8 @@
 // the length of one 3D coordinates (x, y, z) in the coordinate lookup float
 // array
-const ONE_POINT = 3;
-
-/**
-* Swap endianness of 32bit numbers.
-* @param {number} val 32 bit number to be swapped.
-* @return {number} 32bit number in swapped endianness.
-*/
-function swap32(val) {
-  return ((val & 0xFF) << 24) |
-           ((val & 0xFF00) << 8) |
-           ((val >> 8) & 0xFF00) |
-           ((val >> 24) & 0xFF);
-}
+//const ONE_POINT = 3;
+import {ONE_POINT} from './constants.js';
+import {swap32} from './util.js';
 
 /**
  * Constructor of Visualization Class.
