+++ ibr/ibr-sdk/src/constants.js
@@ -1,0 +1,9 @@
+// the length of one 3D coordinates (x, y, z) in the coordinate lookup float
+// array
+export const ONE_POINT = 3;
+// the length of two 3D coordinates (x, y, z) in the coordinate lookup float
+// array
+export const TWO_POINTS = 6;
+
+// separate floor's z-coordinate by 100 unit length
+export const FLOOR_HEIGHT = 300;
